[Chore] Code style improvements and dependency upgrades
Rename exported types with Haruki prefix, modernize error handling
(errors.As), update dependencies (fasthttp v1.70, entgo v0.14.6, etc.),
and apply lint/syntax cleanups across 208 files.

+++ internal/pjsk/drawing/request_dt.go
@@ -100,7 +100,7 @@ func parseDrawingUpdateTime(value any) (int64, bool) {
 	case uint32:
 		return int64(v), true
 	case uint64:
-		if v > uint64(^uint64(0)>>1) {
+		if v > ^uint64(0)>>1 {
 			return 0, false
 		}
 		return int64(v), true
